Made targeting system
Closest target in front of player is prioritized by auto-targeting. 
Player can hold RT or LT to lock on target. Both "Hold" and "Press" targeting is supported.
Variables like targeting range and angle are customizable.
Later additions to this system will include player always facing target when locked on.
IMPORTANT: Player now requires an object with "SCR_ObjectReferenceManager" script to be in the scene.